backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-dotenv
python-jose[cryptography]
python-multipart
httpx
google-genai
faster-whisper
pdfplumber
python-pptx
python-docx
